backend/requirements-current.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==3.7.1
azure-ai-contentsafety==1.0.0
azure-ai-documentintelligence==1.0.2
azure-common==1.1.28
azure-core==1.39.0
azure-cosmos==4.15.0
azure-identity==1.25.3
azure-search-documents==11.6.0
azure-storage-blob==12.28.0
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
ecdsa==0.19.1
fastapi==0.104.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.2
idna==3.11
iniconfig==2.3.0
isodate==0.7.2
jiter==0.13.0
msal==1.35.1
msal-extensions==1.3.1
openai==2.29.0
packaging==26.0
passlib==1.7.4
pluggy==1.6.0
pyasn1==0.6.3
pycparser==3.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
Pygments==2.19.2
PyJWT==2.12.1
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.0.0
python-jose==3.5.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
structlog==25.5.0
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.24.0
watchfiles==1.1.1
websockets==16.0
